Add wallclock time calculation

+++ include/types.h
@@ -5,7 +5,8 @@
 typedef int devhdl_t(devreg_t*); /* Type of a function that takes a device register pointer as argument and
                                            returns int, for device handler routines */
 
-typedef unsigned int cpu_t; /* cpu time type */
+typedef unsigned long long cpu_t; /* cpu time type */
+
 
 typedef enum {PSEUDOCLOCK, TIMESLICE, AGING} timcause_t; /* cause of interval timer interrupt type */
